requirements.txt
pytest
pytest-html
pytest-csv
pyaml
requests